refactor: flatten drip-list-card (delete ...-representational component)
